requirements.txt
Flask
requests
slack
python-dotenv
slackeventsapi
slackclient
pyowm